requirements.txt
opencv-python
opencv-contrib-python
matplotlib
tensorflow==2.*
tensorflow-gpu
numpy
pillow
Cython
lxml